refactor(chat): extract MessageInput into modular sub-components
Split the 1200-line MessageInput into a folder structure:
- MessageInput/index.tsx orchestrates state and props
- MessageInput/ComposerInput.tsx handles TextInput, auto-resize and mention list
- MessageInput/RecordingBar.tsx renders the recording UI (waveform, timer, send/cancel)
- MessageInput/SendOrMicButton.tsx renders the right-side action button
- hooks/useVoiceRecorder.ts encapsulates expo-av recording lifecycle and MIME helpers

The legacy src/components/Chat/MessageInput.tsx is now a thin re-export to keep
existing imports and the buildRecordingOptions test export working unchanged.
All testIDs and accessibility labels are preserved (10/10 tests still green).
